# Use heuristic to select histogram node, avoid rabit call